pyproject.toml
[tool.poetry]
name = "rabbit"
version = "0.1.0"
description = ""
authors = ["noah.Nam <noah85.Nam@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "*"
flask-sqlalchemy = "*"
flask-migrate = "*"
flask-jwt-extended = "*"
psycopg2-binary = "*"
importlib-metadata = "*"
inject = "*"
pydantic = "*"
pypubsub = "*"
flasgger = "*"
greenlet = "*"
gevent = "^20.9.0"
gunicorn = "==20.0.4"
boto3= "1.12.17"
docker = "*"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pytest-xdist = "*"
pytest = "*"
pytest-factoryboy = "*"
faker = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
